refactor: move fromPreference()
to ReviewerBinding
