+++ AnkiDroid/src/main/java/com/ichi2/anki/cardviewer/GestureProcessor.kt
@@ -18,7 +18,7 @@ package com.ichi2.anki.cardviewer
 import android.content.SharedPreferences
 import com.ichi2.anki.reviewer.Binding
 import com.ichi2.anki.reviewer.GestureMapper
-import com.ichi2.anki.reviewer.MappableBinding
+import com.ichi2.anki.reviewer.ReviewerBinding
 
 class GestureProcessor(
     private val processor: ViewerCommand.CommandProcessor?,
@@ -58,7 +58,7 @@ class GestureProcessor(
 
         val associatedCommands = HashMap<Gesture, ViewerCommand>()
         for (command in ViewerCommand.entries) {
-            for (mappableBinding in MappableBinding.fromPreference(preferences, command)) {
+            for (mappableBinding in ReviewerBinding.fromPreference(preferences, command)) {
                 if (mappableBinding.binding is Binding.GestureInput) {
                     associatedCommands[mappableBinding.binding.gesture] = command
                 }
