+++ AnkiDroid/src/test/java/com/ichi2/anki/servicemodel/UpgradeGesturesToControlsTest.kt
@@ -72,7 +72,7 @@ class UpgradeGesturesToControlsTest(
 
         assertThat(prefs.contains(testData.affectedPreferenceKey), equalTo(true))
         assertThat(prefs.contains(testData.unaffectedPreferenceKey), equalTo(false))
-        assertThat("example command should have no defaults", MappableBinding.fromPreference(prefs, command), empty())
+        assertThat("example command should have no defaults", ReviewerBinding.fromPreference(prefs, command), empty())
 
         upgradeAllGestures()
 
@@ -84,7 +84,7 @@ class UpgradeGesturesToControlsTest(
         assertThat("legacy preference removed", prefs.contains(testData.affectedPreferenceKey), equalTo(false))
         assertThat("new preference added", prefs.contains(command.preferenceKey), equalTo(true))
 
-        val fromPreference = MappableBinding.fromPreference(prefs, command)
+        val fromPreference = ReviewerBinding.fromPreference(prefs, command)
         assertThat(fromPreference, hasSize(1))
         val binding = fromPreference.first()
 
@@ -103,7 +103,7 @@ class UpgradeGesturesToControlsTest(
         assertThat(prefs.contains(testData.affectedPreferenceKey), equalTo(true))
         assertThat(prefs.contains(testData.unaffectedPreferenceKey), equalTo(false))
         assertThat("new preference does not exist", prefs.contains(command.preferenceKey), equalTo(false))
-        val previousCommands = MappableBinding.fromPreference(prefs, command)
+        val previousCommands = ReviewerBinding.fromPreference(prefs, command)
         assertThat("example command should have defaults", previousCommands, not(empty()))
 
         upgradeAllGestures()
@@ -116,7 +116,7 @@ class UpgradeGesturesToControlsTest(
         assertThat("legacy preference removed", prefs.contains(testData.affectedPreferenceKey), equalTo(false))
         assertThat("new preference exists", prefs.contains(command.preferenceKey), equalTo(true))
 
-        val currentCommands = MappableBinding.fromPreference(prefs, command)
+        val currentCommands = ReviewerBinding.fromPreference(prefs, command)
         assertThat("a binding was added to '${command.preferenceKey}'", currentCommands, hasSize(previousCommands.size + 1))
 
         // ensure that the order was not changed - the last element is not included in the zip
@@ -143,7 +143,7 @@ class UpgradeGesturesToControlsTest(
         assertThat(prefs.contains(testData.affectedPreferenceKey), equalTo(true))
         assertThat(prefs.contains(testData.unaffectedPreferenceKey), equalTo(false))
         assertThat("new preference exists", prefs.contains(command.preferenceKey), equalTo(true))
-        val previousCommands = MappableBinding.fromPreference(prefs, command)
+        val previousCommands = ReviewerBinding.fromPreference(prefs, command)
         assertThat("example command should have defaults", previousCommands, hasSize(2))
         assertThat(previousCommands.first(), equalTo(testData.binding))
